refactor: review fixes — filter unlisted, role=note, extract sidebar test helpers
- Filter unlisted EN docs from sidebar fallback merge (bug fix)
- Change role="status" to role="note" for static fallback banner (a11y)
- Extract desktopSidebar/waitForSidebarHydration into e2e/sidebar-helpers.ts
- Update both sidebar test files to use shared helpers
