+++ e2e/i18n-sidebar-fallback.spec.ts
@@ -1,5 +1,6 @@
-import { test, expect, type Page, type Locator } from "@playwright/test";
+import { test, expect } from "@playwright/test";
 import { getBasePath } from "./helpers";
+import { desktopSidebar, waitForSidebarHydration } from "./sidebar-helpers";
 
 /**
  * E2E tests for i18n sidebar fallback behavior.
@@ -16,29 +17,6 @@ const BASE = getBasePath();
 
 // Desktop viewport so sidebar is always visible
 test.use({ viewport: { width: 1280, height: 800 } });
-
-function desktopSidebar(page: Page): Locator {
-  return page.locator("#desktop-sidebar");
-}
-
-/** Wait for the React sidebar island to hydrate */
-async function waitForSidebarHydration(page: Page) {
-  const sidebar = desktopSidebar(page);
-  await sidebar.locator("nav").waitFor({ state: "attached" });
-  await page.waitForFunction(
-    () => {
-      const sidebar = document.querySelector("#desktop-sidebar");
-      if (!sidebar) return false;
-      const btn = sidebar.querySelector(
-        'button[aria-label^="Collapse"], button[aria-label^="Expand"]',
-      );
-      if (!btn) return false;
-      return Object.keys(btn).some((k) => k.startsWith("__react"));
-    },
-    null,
-    { timeout: 5000 },
-  );
-}
 
 test.describe("i18n sidebar fallback: JA locale", () => {
   test("JA sidebar shows Sub A and Sub B from EN fallback", async ({
@@ -78,7 +56,7 @@ test.describe("i18n sidebar fallback: JA locale", () => {
     await page.goto(`${BASE}/ja/docs/guides/sub-a/page-1`);
 
     // The fallback notice banner should be visible
-    const banner = page.locator('[role="status"]');
+    const banner = page.locator('[role="note"]');
     await expect(banner).toBeVisible();
     await expect(banner).toContainText(
       "このページはまだ翻訳されていません",
@@ -91,7 +69,7 @@ test.describe("i18n sidebar fallback: JA locale", () => {
     await page.goto(`${BASE}/ja/docs/guides/writing-docs`);
 
     // No fallback banner should exist on a translated page
-    const banner = page.locator('[role="status"]');
+    const banner = page.locator('[role="note"]');
     await expect(banner).toHaveCount(0);
   });
 
@@ -123,7 +101,7 @@ test.describe("i18n sidebar fallback: JA locale", () => {
     await page.waitForURL(/\/ja\/docs\/guides\/sub-a\/page-2/);
 
     // Should show the fallback banner on the new page too
-    const banner = page.locator('[role="status"]');
+    const banner = page.locator('[role="note"]');
     await expect(banner).toBeVisible();
     await expect(banner).toContainText(
       "このページはまだ翻訳されていません",
@@ -148,7 +126,7 @@ test.describe("i18n sidebar fallback: DE locale", () => {
   }) => {
     await page.goto(`${BASE}/de/docs/guides/sub-a/page-1`);
 
-    const banner = page.locator('[role="status"]');
+    const banner = page.locator('[role="note"]');
     await expect(banner).toBeVisible();
     await expect(banner).toContainText(
       "Diese Seite wurde noch nicht übersetzt",
@@ -161,7 +139,7 @@ test.describe("i18n sidebar fallback: EN locale (no fallback)", () => {
     await page.goto(`${BASE}/docs/guides/sub-a/page-1`);
 
     // EN is the default locale — no fallback banner should appear
-    const banner = page.locator('[role="status"]');
+    const banner = page.locator('[role="note"]');
     await expect(banner).toHaveCount(0);
   });
