+++ src/pages/de/docs/[...slug].astro
@@ -175,7 +175,7 @@ const slug = autoIndex ? autoIndex.slug : (entry!.data.slug ?? toRouteSlug(entry
       <DocMetainfo filePath={`${contentDir}/${entry.id}`} locale="de" />
 
       {isFallback && (
-        <div class="mb-vsp-md border border-info/30 bg-info/5 px-hsp-lg py-vsp-sm text-small text-muted rounded" role="status">
+        <div class="mb-vsp-md border border-info/30 bg-info/5 px-hsp-lg py-vsp-sm text-small text-muted rounded" role="note">
           {t("doc.fallbackNotice", "de")}
         </div>
       )}
